separate components by domain, remove unused files
