+++ src/components/common/ResponsiveBlock/index.tsx
