+++ src/components/post/Markdown/renderers/AnchorBlock.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import styled from "styled-components";
 
-import { Color } from "../../../constants/color";
+import { Color } from "../../../../constants/color";
 
 export function AnchorBlock(props: any) {
   return props.href.startsWith("#") || props.href.startsWith("http") ? (
