--- src/components/withPlaceholder/index.tsx
@@ -1,12 +1,0 @@
-import { ComponentType, Suspense } from "react";
-
-export function withPlaceholder<ComponentProps extends {}>(
-  Component: ComponentType<ComponentProps>,
-  Placeholder: ComponentType<{}>
-) {
-  return (props: ComponentProps) => (
-    <Suspense fallback={<Placeholder />}>
-      <Component {...props} />
-    </Suspense>
-  );
-}
