--- src/containers/Main/hooks/useCategories.ts
@@ -1,5 +1,0 @@
-import { useAPIQuery } from "../../../hooks/useAPIQuery";
-
-export function useCategories() {
-  return useAPIQuery("getAllCategories", { count: 12 }, { suspense: true }).data.data;
-}
